fix: moving /prisma into /src/
